+++ wiring/inc/spark_wiring_wifi.h
@@ -121,8 +121,8 @@ public:
         network_off(*this, 0, 0, NULL);
     }
 
-    void listen(void) {
-        network_listen(*this, 0, NULL);
+    void listen(bool begin=true) {
+        network_listen(*this, begin ? 0 : 1, NULL);
     }
 
     bool listening(void) {